requirements.txt
asgiref==3.6.0
astroid==2.15.5
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
openapi-codec==1.3.2
pep8==1.7.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.1
pytz==2023.3
requests==2.30.0
simplejson==3.19.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
